pyproject.toml
[tool.poetry]
name = "red-mascotera"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
packages = [{include = "red_mascotera"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"
django-allauth = "^0.55.2"
djangorestframework-simplejwt = {version = "^5.3.0", extras = ["crypto"]}
drf-yasg = {version = "^1.21.7", extras = ["validation"]}


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.5"
black = "^23.7.0"
pre-commit = "^3.4.0"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
    \.toml
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | \.sh
  | \.idea
  | \.vscode
  | \.ini
  | _build
  | buck-out
  | build
  | dist
  | Dockerfile
  | docker-compose.yml
  | manage.py
)/
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
